backend/requirements.txt
djangorestframework
GDAL==3.7.3
